Move `onScheduleHook` to `Schedulers`
